Update SimpleCV Rectangle Test.py
Not "ready" code, but closer.
Image recognition now shows up on the Pi. A switch flag needs to be added in order to move between a video output and a textual output (Both have been included in the code).

Someone (If not me) needs to look at figuring out a less CPU intensive way to determine geometry in an image. Right now it is not effective enough to qualify loading on to the bot and getting any useful data in a near-real time format.
I would recommend actually looking at how "blobs" are discovered.

Best guess, a pixel-by-pixel scan. This could probably be reduced by performing a 2-1 or 3-1 pixel scan, if not more. The thought behind this is as CPU power decreases, the performance gain vs loss of accuracy would be beneficial over all.

Think about it.