General fixes
Encoder-based unwind now has timer-based function that will give it a 1-second wait prior to unwinding. 
The digital outputs for the Arduino have been commented out. The current position (in the software) of the output pins conflict with the wiring on the robot. A new wiring scheme will have to be constructed.